Customized Email Template
The new file custom-request-new-quote.php  is a customized version of
the request-new-quote.php file.

It contains code to add the Product Image to the email sent to the admin
when a quote request comes through.

The file should be placed inside the active theme with the below path:
<active theme>\quotes-for-wc\emails\

The file should be renamed to request-new-quote.php