Fix #27
The suggestion is to hardcode custom WC text.

Instead, I've removed the code and modified it to return the text that comes through in the function as is.

This will ensure that the plugin modified the button text only if the product has quotes enabled.
At the same time, it will not conflict with any other plugins and ensure future compatibility.